Modifications sur UtilisateursDAO, DAOFactory, fin du développement de UtilisateursDAOJdbcImpl suivant le modèle de AppliDemo.

+++ src/main/java/fr/eni/javaee/ProjetEncheres/bo/Utilisateurs.java
@@ -126,7 +126,7 @@ public class Utilisateurs {
 		this.motDePasse = motDePasse;
 	}
 
-	public long getCredit() {
+	public int getCredit() {
 		return credit;
 	}
